Rid of the OO wrapper
Stays close to the C-api, and directly exports the binding's functions
and annotates their types.
