+++ src/modules/VL53L5CX/package.json
@@ -15,5 +15,9 @@
   ],
   "author": "",
   "license": "MIT",
-  "gypfile": true
+  "gypfile": true,
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
+  }
 }
